# cmd/api-server/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  api.ApplyNATSRequest:
    properties:
      dry_run:
        type: boolean
      manifest:
        $ref: '#/definitions/natsadmin.Manifest'
    type: object
  api.Candidate:
    properties:
      batch_id:
        type: string
      description:
        type: string
      discovered_at:
        type: string
      id:
        type: string
      ingestion_method:
        type: string
      published_at:
        type: string
      source_abbr:
        type: string
      title:
        type: string
      trace_id:
        type: string
      url:
        type: string
    type: object
...
  api.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  api.ListCandidatesResponse:
    properties:
      count:
        type: integer
      items:
        items:
          $ref: '#/definitions/api.Candidate'
        type: array
      limit:
        type: integer
      offset:
        type: integer
    type: object
  api.NATSTestMessageRequest:
    properties:
      subject:
        type: string
    type: object
...
  api.PauseNATSConsumerRequest:
    properties:
      reason:
        type: string
      until:
        type: string
    type: object
...
  natsadmin.ApplyReport:
    properties:
      created:
        items:
          type: string
        type: array
      unchanged:
        items:
          type: string
        type: array
      updated:
        items:
          type: string
        type: array
    type: object
  natsadmin.ConsumerManifest:
    properties:
      ack_policy:
        type: string
      ack_wait:
        type: string
      deliver_policy:
        type: string
      filter_subject:
        type: string
      max_ack_pending:
        type: integer
      max_deliver:
        type: integer
      name:
        type: string
      queue_group:
        type: string
      stream:
        type: string
    type: object
  natsadmin.Manifest:
    properties:
      consumers:
        items:
          $ref: '#/definitions/natsadmin.ConsumerManifest'
        type: array
      streams:
        items:
          $ref: '#/definitions/natsadmin.StreamManifest'
        type: array
      version:
        type: integer
    type: object
  natsadmin.PauseResult:
    properties:
      consumer:
        type: string
      pause_until:
        type: string
      paused:
        type: boolean
      reason:
        type: string
      stream:
        type: string
    type: object
  natsadmin.StreamManifest:
    properties:
      allow_test_publish:
        type: boolean
      max_age:
        type: string
      max_messages:
        type: integer
      name:
        type: string
      replicas:
        type: integer
      retention:
        type: string
      storage:
        type: string
      subjects:
        items:
          type: string
        type: array
    type: object
  natsadmin.TestMessageResult:
    properties:
      sequence:
        type: integer
      stream:
        type: string
      subject:
        type: string
    type: object
  natsdiag.ConsumerInfo:
    properties:
      ack_floor_stream_sequence:
        type: integer
      ack_pending:
        type: integer
      delivered_stream_sequence:
        type: integer
      filter_subject:
        type: string
      last_active:
        type: string
      name:
        type: string
      pending:
        type: integer
      redelivered:
        type: integer
      stream:
        type: string
      waiting:
        type: integer
    type: object
  natsdiag.Snapshot:
    properties:
      collected_at:
        type: string
      rtt:
        type: integer
      streams:
        items:
          $ref: '#/definitions/natsdiag.StreamInfo'
        type: array
    type: object
  natsdiag.StreamInfo:
    properties:
      bytes:
        type: integer
      consumers:
        items:
          $ref: '#/definitions/natsdiag.ConsumerInfo'
        type: array
      first_sequence:
        type: integer
      last_sequence:
        type: integer
      messages:
        type: integer
      name:
        type: string
      subject_counts:
        additionalProperties:
          format: int64
          type: integer
        type: object
      subjects:
        items:
          type: string
        type: array
    type: object
...
info:
  contact: {}
  description: User-facing read API for Prism candidate/content data.
  title: Prism API
  version: "0.1"
paths:
...
  /admin/candidates:
    get:
      parameters:
      - description: Keyword matched against title/description (ILIKE)
        in: query
        name: q
        type: string
      - description: Filter by source abbreviation (e.g. dpp, tpp, yahoo)
        in: query
        name: source_abbr
        type: string
      - description: Lower bound on published_at/discovered_at (RFC3339)
        in: query
        name: since
        type: string
      - description: Upper bound on published_at/discovered_at (RFC3339)
        in: query
        name: until
        type: string
      - description: Page size (default 50, max 200)
        in: query
        name: limit
        type: integer
      - description: Pagination offset (default 0)
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.ListCandidatesResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: List candidate article briefs
      tags:
      - candidates
...
  /admin/nats:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/natsdiag.Snapshot'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Inspect NATS JetStream state
      tags:
      - admin
  /admin/nats/apply:
    post:
      consumes:
      - application/json
      parameters:
      - description: JetStream manifest
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.ApplyNATSRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/natsadmin.ApplyReport'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Apply a JetStream manifest
      tags:
      - admin
  /admin/nats/streams/{stream}:
    delete:
      parameters:
      - description: Stream name
        in: path
        name: stream
        required: true
        type: string
      - description: Stream name confirmation
        in: query
        name: confirm
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "502":
          description: Bad Gateway
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Delete a JetStream stream
      tags:
      - admin
  /admin/nats/streams/{stream}/consumers/{consumer}:
    delete:
      parameters:
      - description: Stream name
        in: path
        name: stream
        required: true
        type: string
      - description: Consumer name
        in: path
        name: consumer
        required: true
        type: string
      - description: Consumer name confirmation
        in: query
        name: confirm
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "502":
          description: Bad Gateway
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Delete a JetStream consumer
      tags:
      - admin
  /admin/nats/streams/{stream}/consumers/{consumer}/pause:
    post:
      consumes:
      - application/json
      parameters:
      - description: Stream name
        in: path
        name: stream
        required: true
        type: string
      - description: Consumer name
        in: path
        name: consumer
        required: true
        type: string
      - description: Pause deadline
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.PauseNATSConsumerRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/natsadmin.PauseResult'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "502":
          description: Bad Gateway
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Pause a JetStream consumer
      tags:
      - admin
  /admin/nats/streams/{stream}/consumers/{consumer}/resume:
    post:
      parameters:
      - description: Stream name
        in: path
        name: stream
        required: true
        type: string
      - description: Consumer name
        in: path
        name: consumer
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/natsadmin.PauseResult'
        "502":
          description: Bad Gateway
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Resume a JetStream consumer
      tags:
      - admin
  /admin/nats/streams/{stream}/purge:
    post:
      parameters:
      - description: Stream name
        in: path
        name: stream
        required: true
        type: string
      - description: Stream name confirmation
        in: query
        name: confirm
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "502":
          description: Bad Gateway
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Purge a JetStream stream
      tags:
      - admin
  /admin/nats/test-message:
    post:
      consumes:
      - application/json
      parameters:
      - description: Test subject
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.NATSTestMessageRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/natsadmin.TestMessageResult'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "502":
          description: Bad Gateway
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Publish a restricted JetStream test message
      tags:
      - admin
...
  /candidates:
    get:
      parameters:
      - description: Keyword matched against title/description (ILIKE)
        in: query
        name: q
        type: string
      - description: Filter by source abbreviation (e.g. dpp, tpp, yahoo)
        in: query
        name: source_abbr
        type: string
      - description: Lower bound on published_at/discovered_at (RFC3339)
        in: query
        name: since
        type: string
      - description: Upper bound on published_at/discovered_at (RFC3339)
        in: query
        name: until
        type: string
      - description: Page size (default 50, max 200)
        in: query
        name: limit
        type: integer
      - description: Pagination offset (default 0)
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.ListCandidatesResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: List candidate article briefs
      tags:
      - candidates
...
swagger: "2.0"